% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IN2R.R
\name{IN2R}
\alias{IN2R}
\title{Fetch parameters from CAT-ROMA file}
\usage{
IN2R(f = NULL)
}
\arguments{
\item{f}{NULL, file location, or chooser if left NULL.}
}
\value{
MCAT_itembank
}
\description{
Fetch parameters from CAT-ROMA file
}

